backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.35
Werkzeug==3.0.1
openai==1.99.5
requests==2.31.0
PyJWT==2.8.0
python-dotenv==1.0.0